requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
openai
anthropic
qdrant-client
langgraph
python-multipart
